Refactor transcription and remove unused classes
Removed TranscriptionProcessor and KeywordReplacement classes, and updated project references accordingly. Added new TranscriptionResult and TranscriptionSegment classes, and refactored related code to use these. Improved microphone selection logic in MainRibbon and updated UML diagram for TranscriptionService.
